Fixed CS0436 when used in multiple projects and hidden the attribute from the build output
